pyapps/requirements.txt
# requirements for pybot
python-telegram-bot
pyyaml
icecream
flask